+++ modules/profiles/base.nix
@@ -8,6 +8,8 @@ in
   # Prefer our base config. This also prevents adding ZFS to `boot.supportedFilesystems` without forcing it.
   disabledModules = [ trilby.nixpkgs.nixosModules.profiles.base ];
 
+  environment.pathsToLink = ["/share/X11"];
+
   environment.systemPackages = with pkgs; [
     expect
     file
@@ -19,6 +21,7 @@ in
     trilby-cli
     unzip
     wget
+    xkeyboard-config
     zip
   ];
 }